Move estimateGasDataForNoSignature to test helpers
