+++ universal-login-contracts/test/contracts/Executor/Executor-void.ts
@@ -3,8 +3,8 @@ import {createMockProvider, deployContract, getWallets} from 'ethereum-waffle';
 import Executor from '../../../build/TestableExecutor.json';
 import {constants, Contract} from 'ethers';
 import {transferMessage} from '../../helpers/ExampleMessages';
-import {getExecutionArgs} from '../../helpers/argumentsEncoding';
-import {estimateGasLimit, estimateGasDataForNoSignature} from '../../../lib/estimateGas';
+import {getExecutionArgs, estimateGasDataForNoSignature} from '../../helpers/argumentsEncoding';
+import {estimateGasLimit} from '../../../lib/estimateGas';
 
 
 describe('Void Executor', () => {
