+++ universal-login-contracts/test/helpers/argumentsEncoding.ts
@@ -2,6 +2,7 @@ import {UnsignedMessage} from '@universal-login/commons';
 import {Interface, BigNumberish, Arrayish, arrayify, solidityKeccak256} from 'ethers/utils';
 import {Wallet, utils, Contract} from 'ethers';
 import WalletProxy from '../../build/WalletProxy.json';
+import {estimateGasDataFromSignedMessage} from '../../lib/estimateGas';
 
 export const switchENSNameInInitializeArgs = (initializeArgs: string[], label: string, domain = 'mylogin.eth') => {
   const ensName = `${label}.${domain}`;
@@ -47,3 +48,5 @@ export const setupUpdateMessage = async (proxyAsWalletContract: Contract, newWal
     gasToken: '0x0000000000000000000000000000000000000000'
   };
 };
+
+export const estimateGasDataForNoSignature = (unsignedMessage: UnsignedMessage) => estimateGasDataFromSignedMessage({...unsignedMessage, signature: '0x0'});
